+++ src/App.tsx
@@ -1,16 +1,14 @@
-import React, { FC } from "react";
+import React from "react";
 import { Provider } from "react-redux";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 import ReactNotification from "react-notifications-component";
 import "react-notifications-component/dist/theme.css";
 import { GamePage, HomePage, NotFoundPage, LoginPage } from "@/pages";
+import { Widgets, GenerationWidget, BalanceWidget } from "@/modules";
 import { store } from "./store";
 import { PageContent, PageWrapper, SideColumn } from "./AppComponents";
-import { Widgets } from "./modules";
-import { GenerationWidget } from "./modules/Statistics";
-import { BalanceWidget } from "./modules/Money";
 
-const App: FC = () => {
+const App = () => {
   return (
     <Provider store={store}>
       <ReactNotification />
@@ -19,7 +17,7 @@ const App: FC = () => {
           <BrowserRouter>
             <Switch>
               <Route path="/game" component={GamePage} />
-              <Route path="/login" exact component={LoginPage} />
+              <Route path="/login" component={LoginPage} />
               <Route path="/" exact component={HomePage} />
               <Route path="*" component={NotFoundPage} />
             </Switch>
