+++ src/modules/Account/index.ts
@@ -1,4 +1,5 @@
 export { accountSlice } from "./reducer";
 export { AccessChecker } from "./AccessChecker";
+export { LoginForm } from "./LoginForm";
 export { AccountWidget } from "./AccountWidget";
 export { accountSaga } from "./saga";
