+++ src/modules/index.ts
@@ -1,7 +1,12 @@
 export { Widgets } from "./Widgets";
-
-// TODO: add...
-
+export { AccessChecker, AccountWidget, LoginForm } from "./Account";
+export { BetWindow } from "./Bets";
+export { BalanceWidget } from "./Money";
+export {
+  FilledPercentStatistics,
+  GenerationWidget,
+  HeightStatistics,
+} from "./Statistics";
 export {
   GameOfLife,
   SizeWidget,
