+++ src/pages/LoginPage.tsx
@@ -1,7 +1,6 @@
 import React from "react";
-import { LoginForm } from "@/modules/Account/LoginForm";
+import { LoginForm, AccessChecker } from "@/modules";
 import { RelativeWrapper, AbsoluteCenter } from "./styled";
-import { AccessChecker } from "@/modules/Account/AccessChecker";
 
 export const LoginPage = () => (
   <AccessChecker onlyUnauthorized redirectPath="/">
