+++ src/pages/NotFoundPage.tsx
@@ -1,3 +1,4 @@
 import React from "react";
+import { H1 } from "./styled";
 
-export const NotFoundPage = () => <div>NotFoundPage</div>;
+export const NotFoundPage = () => <H1>NotFoundPage</H1>;
